Moved ObjectDetectorBackend to separate file
